pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["oguzok2012"]


[tool.poetry.dependencies]
python = "^3.10"
selenium = "4.42.0"
setuptools = "*"
undetected-chromedriver = "3.5.5"
beautifulsoup4 = "4.14.3"
python-dotenv = "^1.0.1"
requests = "^2.31.0"